Added open/closed points, fixed f dist at 0
Added controls for open and closed points and separated these controls
for points on the graph and off the graph of a given equation.  Also
fixed evaluation issue in sum of multiple integrals/areas under the
curve.  Fixed issue in parsing x and y values from string point (e.g.
'(0, 1)') so that it identifies values with decimals.